% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\name{aamutation}
\alias{aamutation}
\alias{aamutation,Variant-method}
\title{Extract amino acid mutation position}
\usage{
aamutation(variant)

\S4method{aamutation}{Variant}(variant)
}
\arguments{
\item{variant}{Variant object}
}
\value{
Amino acid position as character (or "-" if non-coding)
}
\description{
Gets the affected amino acid position from HGVSp notation
}
\section{Methods (by class)}{
\itemize{
\item \code{aamutation(Variant)}: Default implementation for Variant objects

}}

\examples{
# Create a GRanges object for position
pos <- GRanges(seqnames = "chr1", ranges = IRanges(start = 123456, width = 1), strand = "+")

#create an snv
snv <- SNV(
  variantID = "rs123",
  position = pos,
  refAllele = "A",
  altAllele = "T",
  geneSymbol = "BRCA1",
  geneType = "protein-coding",
  effect = "missense",
  clinicalRelevance = "Pathogenic",
  ensemblID = "ENST00000357654",
  hgvsp = "p.Arg175His"
)

aamutation(snv)
}

