% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\name{altAllele}
\alias{altAllele}
\alias{altAllele,Variant-method}
\alias{altAlleleSet}
\alias{altAlleleSet,Variant-method}
\title{Get or set alternative allele}
\usage{
altAllele(x)

\S4method{altAllele}{Variant}(x)

altAlleleSet(x, value)

\S4method{altAlleleSet}{Variant}(x, value)
}
\arguments{
\item{x}{A Variant object}

\item{value}{A character string representing the new alternative allele (for setter)}
}
\value{
The alternative allele (getter) or updated object (setter)
}
\description{
Get or set alternative allele
}


\examples{
pos <- GRanges(seqnames = "chr1", ranges = IRanges(start = 123456, width = 1), strand = "+")

#create an snv
snv <- SNV(
  variantID = "rs123",
  position = pos,
  refAllele = "A",
  altAllele = "T",
  geneSymbol = "BRCA1",
  geneType = "protein-coding",
  effect = "missense",
  clinicalRelevance = "Pathogenic",
  ensemblID = "ENST00000357654",
  hgvsp = "p.Arg175His"
)

altAllele(snv)
}

