% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_protein.R
\name{amino_acid_code_1_to_3}
\alias{amino_acid_code_1_to_3}
\title{Convert snv table with protein column from one letter to three letter amino acid code}
\usage{
amino_acid_code_1_to_3(snv)
}
\arguments{
\item{snv}{}
}
\description{
Convert snv table with protein column from one letter to three letter amino acid code
}
