% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_protein.R
\name{amino_acid_code_3_to_1}
\alias{amino_acid_code_3_to_1}
\title{Convert snv table with protein column from three letter to single letter amino acid code}
\usage{
amino_acid_code_3_to_1(snv)
}
\arguments{
\item{snv}{}
}
\description{
Convert snv table with protein column from three letter to single letter amino acid code
}
