% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_filepaths.R
\name{annotation_filepaths}
\alias{annotation_filepaths}
\title{Generate list of filepaths from annotation folder}
\usage{
annotation_filepaths(output_path)
}
\arguments{
\item{output_path}{}
}
\description{
Generate list of filepaths from annotation folder
}
