% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNOMAD_functions.R
\name{attach_gnomad_MAF}
\alias{attach_gnomad_MAF}
\title{Assign identified MAF to tsnv table}
\usage{
attach_gnomad_MAF(snv_table, gnomad)
}
\arguments{
\item{snv_table}{}
}
\description{
Assign identified MAF to tsnv table
}
