% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRAF_variant_class_module.R
\name{BRAF_class_retrieve_table}
\alias{BRAF_class_retrieve_table}
\title{BRAF class table retrieve}
\usage{
BRAF_class_retrieve_table(variant_table, mp_vars_table)
}
\arguments{
\item{mp_vars_table}{}
}
\description{
BRAF class table retrieve
}
