% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRAF_variant_class_module.R
\name{BRAF_variant_check}
\alias{BRAF_variant_check}
\title{Check MP_variant_table for variant interpretations}
\usage{
BRAF_variant_check(rowid, genestr, protein, mp_vars_table)
}
\arguments{
\item{mp_vars_table}{}
}
\description{
Check MP_variant_table for variant interpretations
}
