% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_counter.R
\name{cancerHotspot_add_columns_to_snv}
\alias{cancerHotspot_add_columns_to_snv}
\title{Assign identified MAF to tsnv table}
\usage{
cancerHotspot_add_columns_to_snv(snv_table)
}
\arguments{
\item{snv_table}{}
}
\description{
Assign identified MAF to tsnv table
}
