% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_functions.R
\name{cancerHotspot_add_mutation_values}
\alias{cancerHotspot_add_mutation_values}
\title{Assign identified MAF to tsnv table}
\usage{
cancerHotspot_add_mutation_values(snv_table, cancerHotspots)
}
\arguments{
\item{snv_table}{}
}
\description{
Assign identified MAF to tsnv table
}
