% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_functions.R
\name{cancerHotspot_adjust_protein_column}
\alias{cancerHotspot_adjust_protein_column}
\title{Adjust - if needed -- protein column to be single amino acid letter encoding}
\usage{
cancerHotspot_adjust_protein_column(snv_table)
}
\arguments{
\item{snv_table}{}
}
\description{
Adjust - if needed -- protein column to be single amino acid letter encoding
}
