% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_counter.R
\name{cancerHotspot_info}
\alias{cancerHotspot_info}
\title{Run cancerHotspot data extraction pipeline}
\usage{
cancerHotspot_info(snv_table, cancerHotspots)
}
\arguments{
\item{cancerHotspots}{}
}
\description{
Run cancerHotspot data extraction pipeline
}
