% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_counter.R
\name{cancerHotspots_mutation_positionCount}
\alias{cancerHotspots_mutation_positionCount}
\title{Retrieve if variant codon position is covered in cancerHotspots v2}
\usage{
cancerHotspots_mutation_positionCount(
  geneName,
  mutation_position,
  cancerHotspots
)
}
\arguments{
\item{snv_position}{}
}
\description{
Retrieve if variant codon position is covered in cancerHotspots v2
}
\examples{
Retrieves No. of total entries in CANCER_HOTSPOTS
KRAS p.G12C retrieves n_total == 2175
}
