% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_counter.R
\name{cancerHotspots_mutationCount}
\alias{cancerHotspots_mutationCount}
\title{Title}
\usage{
cancerHotspots_mutationCount(
  geneName,
  mutation_position,
  mut_AA,
  cancerHotspots
)
}
\arguments{
\item{cancerHotspots}{}
}
\description{
Title
}
