% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTSGpos.R
\name{canonical_splicesite}
\alias{canonical_splicesite}
\title{Determine if splice is at +/- 1 or 2 position relative to coding position}
\usage{
canonical_splicesite(numberlist)
}
\arguments{
\item{numberlist}{}
}
\description{
Determine if splice is at +/- 1 or 2 position relative to coding position
}
