% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLINVAR_functions.R
\name{clinvar_check_coding}
\alias{clinvar_check_coding}
\title{Check coding string for presence in clinvar_fil}
\usage{
clinvar_check_coding(codingstr, clinvar_fil)
}
\arguments{
\item{clinvar_fil}{}
}
\description{
Check coding string for presence in clinvar_fil
}
