% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLINVAR_functions.R
\name{clinvar_check_gene}
\alias{clinvar_check_gene}
\title{Check gene string for presence in clinvar_fil}
\usage{
clinvar_check_gene(genestr, clinvar_fil)
}
\arguments{
\item{clinvar_fil}{}
}
\description{
Check gene string for presence in clinvar_fil
}
