% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLINVAR_functions.R
\name{clinvar_check_protein}
\alias{clinvar_check_protein}
\title{Check protein string for presence in clinvar_fil}
\usage{
clinvar_check_protein(proteinstr, clinvar_fil)
}
\arguments{
\item{clinvar_fil}{}
}
\description{
Check protein string for presence in clinvar_fil
}
