% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLINVAR_functions.R
\name{clinvar_filtering}
\alias{clinvar_filtering}
\title{Filter CLINVAR table based on gene, protein coding first.
If coding and protein are  given, but coding does not match, then search for proteinstr only}
\usage{
clinvar_filtering(genestr, codingstr, proteinstr, clinvar)
}
\arguments{
\item{clinvar}{}
}
\description{
Filter CLINVAR table based on gene, protein coding first.
If coding and protein are  given, but coding does not match, then search for proteinstr only
}
