% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLINVAR_functions.R
\name{ClinVar_function_call}
\alias{ClinVar_function_call}
\title{Calls function to process variant table and check variants for entries in Clinvar table}
\usage{
ClinVar_function_call(asnv, clinvar)
}
\arguments{
\item{clinvar}{}
}
\description{
Calls function to process variant table and check variants for entries in Clinvar table
}
