% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLINVAR_functions.R
\name{clinvar_significance}
\alias{clinvar_significance}
\title{Return Clinical significance entry from Clinvar table}
\usage{
clinvar_significance(clinvar_hit)
}
\arguments{
\item{clinvar_hit}{}
}
\description{
Return Clinical significance entry from Clinvar table
}
