% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLINVAR_functions.R
\name{clinvar_variantID}
\alias{clinvar_variantID}
\title{Return Variation ID entry from Clinvar table}
\usage{
clinvar_variantID(clinvar_hit)
}
\arguments{
\item{clinvar_hit}{}
}
\description{
Return Variation ID entry from Clinvar table
}
