\name{codeAlleleObservations}
\alias{codeAlleleObservations}
\title{Look up integer vals for observed bases...}
\usage{codeAlleleObservations(allele.options, observed)
codeAlleleObservations(allele.options, observed)
codeAlleleObservations(allele.options, observed)
codeAlleleObservations(allele.options, observed)
codeAlleleObservations(allele.options, observed)
codeAlleleObservations(allele.options, observed)
}
\description{Look up integer vals for observed bases}
\details{.Call version of codeAlleleObservations.
Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Return column
for observed alleles in one sample.\code{codeAlleleObservations}: .Call version of codeAlleleObservations.
Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Return column
for observed alleles in one sample.


\code{codeAlleleObservations}: .Call version of codeAlleleObservations.
Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Return column
for observed alleles in one sample.


\code{codeAlleleObservations}: .Call version of codeAlleleObservations.
Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Return column
for observed alleles in one sample.


\code{codeAlleleObservations}: .Call version of codeAlleleObservations.
Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Return column
for observed alleles in one sample.


\code{codeAlleleObservations}: .Call version of codeAlleleObservations.
Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Return column
for observed alleles in one sample.


}
\value{character, like c("0/1","0/0") for reference, alt option 1, then homref, etc.\code{codeAlleleObservations}: character, like c("0/1","0/0") for reference, alt option 1, then homref, etc.


\code{codeAlleleObservations}: character, like c("0/1","0/0") for reference, alt option 1, then homref, etc.


\code{codeAlleleObservations}: character, like c("0/1","0/0") for reference, alt option 1, then homref, etc.


\code{codeAlleleObservations}: character, like c("0/1","0/0") for reference, alt option 1, then homref, etc.


\code{codeAlleleObservations}: character, like c("0/1","0/0") for reference, alt option 1, then homref, etc.


}
\author{Peter M. Haverty \email{phaverty@gene.com}}
\arguments{\item{allele.options}{list from codeAlleleOptions}
\item{observed}{character, like c("AA","AC","CA") from genotypes for a given sample}
}

\alias{codeAlleleObservations}

\alias{codeAlleleObservations}

\alias{codeAlleleObservations}

\alias{codeAlleleObservations}

\alias{codeAlleleObservations}

