\name{codeAlleleOptions}
\alias{codeAlleleOptions}
\title{Make lookup for alleles...}
\usage{codeAlleleOptions(refbase, alleles)
codeAlleleOptions(refbase, alleles)
codeAlleleOptions(refbase, alleles)
codeAlleleOptions(refbase, alleles)
codeAlleleOptions(refbase, alleles)
codeAlleleOptions(refbase, alleles)
}
\description{Make lookup for alleles}
\details{Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Update list to make reference
first position. This lookup will be used to encode each new sample.\code{codeAlleleOptions}: Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Update list to make reference
first position. This lookup will be used to encode each new sample.


\code{codeAlleleOptions}: Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Update list to make reference
first position. This lookup will be used to encode each new sample.


\code{codeAlleleOptions}: Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Update list to make reference
first position. This lookup will be used to encode each new sample.


\code{codeAlleleOptions}: Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Update list to make reference
first position. This lookup will be used to encode each new sample.


\code{codeAlleleOptions}: Take references bases (vector of A, C, G, or T) and alleles (list of
vectors with allowed letters (i.e list( c("A","C") )). Update list to make reference
first position. This lookup will be used to encode each new sample.


}
\value{character, like c("0/1","0/0")\code{codeAlleleOptions}: character, like c("0/1","0/0")


\code{codeAlleleOptions}: character, like c("0/1","0/0")


\code{codeAlleleOptions}: character, like c("0/1","0/0")


\code{codeAlleleOptions}: character, like c("0/1","0/0")


\code{codeAlleleOptions}: character, like c("0/1","0/0")


}
\author{Peter M. Haverty \email{phaverty@gene.com}}
\arguments{\item{refbase}{character, one letter per position specifying base in reference genome}
\item{alleles}{list, one character vector per element in refbase, letters allowed at
this position.}
}

\alias{codeAlleleOptions}

\alias{codeAlleleOptions}

\alias{codeAlleleOptions}

\alias{codeAlleleOptions}

\alias{codeAlleleOptions}

