% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSMIC_functions.R
\name{COSMIC_function_call}
\alias{COSMIC_function_call}
\title{Count variant occurences per tissue in COSMIC db}
\usage{
COSMIC_function_call(snv_table, sql_con_tbl)
}
\arguments{
\item{sql_con_tbl}{}
}
\description{
Count variant occurences per tissue in COSMIC db
}
