% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\docType{class}
\name{Deletion-class}
\alias{Deletion-class}
\title{Deletion class}
\description{
Class for deletion mutations
}
\section{Slots}{

\describe{
\item{\code{deletedSeq}}{The deleted sequence}
}}


\examples{
# Create a GRanges object for a range
pos_del <- GRanges("chr5", IRanges(500, 502))

#create a deletion
del <- Deletion(
  variantID = "del1",
  position = pos_del,
  refAllele = "TGC",
  geneSymbol = "PTEN",
  geneType = "protein-coding",
  effect = "frameshift",
  clinicalRelevance = "Pathogenic",
  ensemblID = "ENST00000371953",
  hgvsp = "p.Asp117fs"
)

del

}

