% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\name{Deletion}
\alias{Deletion}
\title{Deletion constructor}
\usage{
Deletion(variantID, position, refAllele, ...)
}
\arguments{
\item{variantID}{Variant identifier}

\item{position}{GRanges object (deleted region)}

\item{refAllele}{Reference sequence being deleted}

\item{...}{Additional Variant slots}
}
\value{
Deletion object
}
\description{
Creates a Deletion object
}


\examples{
# Create a GRanges object for a range
pos_del <- GRanges("chr5", IRanges(500, 502))

#create a deletion
del <- Deletion(
  variantID = "del1",
  position = pos_del,
  refAllele = "TGC",
  geneSymbol = "PTEN",
  geneType = "protein-coding",
  effect = "frameshift",
  clinicalRelevance = "Pathogenic",
  ensemblID = "ENST00000371953",
  hgvsp = "p.Asp117fs"
)

del

}

