\name{VariantAnnotation-deprecated}
\alias{VariantAnnotation-deprecated}
\title{Deprecated Functions in Package VariantAnnotation}

\description{
These functions are provided for compatibility with older versions of R only, and may be defunct as soon as the next release.
}

\details{
  \link{MatrixToSnpMatrix-deprecated}
}