% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_CNVs.R
\name{determine_cnv_status}
\alias{determine_cnv_status}
\title{Determine CNV status of CNVs}
\usage{
determine_cnv_status(perc_05, perc_95)
}
\arguments{
\item{perc_95}{}
}
\description{
Determine CNV status of CNVs
}
