% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\docType{class}
\name{DNV-class}
\alias{DNV-class}
\title{Double Nucleotide Variant (DNV) class}
\description{
Class for dinucleotide substitutions (e.g., AG>TC)
}


\examples{
# Create a GRanges object for position
pos2 <- GRanges(seqnames = "chr2", ranges = IRanges(start = 200, width = 2))

#create a DNV
dnv <- DNV(
  variantID = "dnv1",
  position = pos2,
  refAllele = "AG",
  altAllele = "TC",
  geneSymbol = "TP53",
  geneType = "protein-coding",
  effect = "nonsense",
  clinicalRelevance = "Likely pathogenic",
  ensemblID = "ENST00000269305",
  hgvsp = "p.Trp53*"
)

dnv
}

