% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\name{DNV}
\alias{DNV}
\title{DNV constructor}
\usage{
DNV(variantID, position, refAllele, altAllele, ...)
}
\arguments{
\item{variantID}{Variant identifier}

\item{position}{GRanges object (must span 2 adjacent nucleotides)}

\item{refAllele}{Reference dinucleotide}

\item{altAllele}{Alternate dinucleotide}

\item{...}{Additional Variant slots}
}
\value{
DNV object
}
\description{
Creates a DNV (dinucleotide variant) object
}


\examples{
# Create a GRanges object for position
pos2 <- GRanges(seqnames = "chr2", ranges = IRanges(start = 200, width = 2))

#create a DNV
dnv <- DNV(
  variantID = "dnv1",
  position = pos2,
  refAllele = "AG",
  altAllele = "TC",
  geneSymbol = "TP53",
  geneType = "protein-coding",
  effect = "nonsense",
  clinicalRelevance = "Likely pathogenic",
  ensemblID = "ENST00000269305",
  hgvsp = "p.Trp53*"
)

dnv
}

