% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\name{effect}
\alias{effect}
\alias{effect,Variant-method}
\title{Get variant effect}
\usage{
effect(x)

\S4method{effect}{Variant}(x)
}
\arguments{
\item{x}{Variant object}
}
\value{
Variant effect (e.g., "missense")
}
\description{
Get variant effect
}
\section{Methods (by class)}{
\itemize{
\item \code{effect(Variant)}: Get variant effect for any Variant

}}

\examples{
# Create a dummy Variant object (assuming class definition exists)
pos <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, width=1), strand = "+")
snv <- SNV(
  variantID = "rs1",
  position = pos,
  refAllele = "A",
  altAllele = "T",
  geneSymbol = "GENE1",
  geneType = "protein-coding",
  effect = "missense",
  clinicalRelevance = "Pathogenic",
  ensemblID = "ENST0000001",
  hgvsp = "p.Val1Gly"
)

effect(snv)
}

