% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_snv_AA_position.R
\name{extract_number_from_alphanumeric_string}
\alias{extract_number_from_alphanumeric_string}
\title{Extract position of SNVs}
\usage{
extract_number_from_alphanumeric_string(alphanumstring)
}
\arguments{
\item{clinvar_ready_AA}{}
}
\value{
SNV amino-acid position
}
\description{
Extract position of SNVs
}
