% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTSGpos.R
\name{extract_splitesite}
\alias{extract_splitesite}
\title{Extracts digits following +/- annotation in codingstring
Importantly, only extracts first hit}
\usage{
extract_splitesite(codingstring)
}
\arguments{
\item{codingstring}{}
}
\description{
Extracts digits following +/- annotation in codingstring
Importantly, only extracts first hit
}
