\name{fixNA}
\alias{fixNA}
\title{Code NA as ".}
\usage{fixNA(x)
fixNA(x)
fixNA(x)
fixNA(x)
fixNA(x)
fixNA(x)
}
\description{Code NA as "."}
\details{Code NA as "."\code{fixNA}: Code NA as "."


\code{fixNA}: Code NA as "."


\code{fixNA}: Code NA as "."


\code{fixNA}: Code NA as "."


\code{fixNA}: Code NA as "."


}
\value{vector\code{fixNA}: vector


\code{fixNA}: vector


\code{fixNA}: vector


\code{fixNA}: vector


\code{fixNA}: vector


}
\author{Peter M. Haverty \email{phaverty@gene.com}}
\arguments{\item{x}{vector}
}

\alias{fixNA}

\alias{fixNA}

\alias{fixNA}

\alias{fixNA}

\alias{fixNA}

