% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_CNVs.R
\name{gene_cnv_effect}
\alias{gene_cnv_effect}
\title{Compare CNV status with CNV effect in cnv lut}
\usage{
gene_cnv_effect(gene, cnv_status, cnv_lut)
}
\arguments{
\item{cnv_lut}{}
}
\description{
Compare CNV status with CNV effect in cnv lut
}
