\name{genotypesFromPinfo}
\alias{genotypesFromPinfo}
\title{Parse probe info file to get nucleotide options...}
\usage{genotypesFromPinfo(pinfo)
genotypesFromPinfo(pinfo)
genotypesFromPinfo(pinfo)
genotypesFromPinfo(pinfo)
genotypesFromPinfo(pinfo)
genotypesFromPinfo(pinfo)
}
\description{Parse probe info file to get nucleotide options}
\details{Parse probe info file to get nucleotide options on top strand.\code{genotypesFromPinfo}: Parse probe info file to get nucleotide options on top strand.


\code{genotypesFromPinfo}: Parse probe info file to get nucleotide options on top strand.


\code{genotypesFromPinfo}: Parse probe info file to get nucleotide options on top strand.


\code{genotypesFromPinfo}: Parse probe info file to get nucleotide options on top strand.


\code{genotypesFromPinfo}: Parse probe info file to get nucleotide options on top strand.


}
\value{list of characters, e.g. list( c("A","G"), c("C","A"))\code{genotypesFromPinfo}: list of characters, e.g. list( c("A","G"), c("C","A"))


\code{genotypesFromPinfo}: list of characters, e.g. list( c("A","G"), c("C","A"))


\code{genotypesFromPinfo}: list of characters, e.g. list( c("A","G"), c("C","A"))


\code{genotypesFromPinfo}: list of characters, e.g. list( c("A","G"), c("C","A"))


\code{genotypesFromPinfo}: list of characters, e.g. list( c("A","G"), c("C","A"))


}
\author{Peter M. Haverty \email{phaverty@gene.com}}
\arguments{\item{pinfo}{data.frame, as read by load.pinfo()}
}

\alias{genotypesFromPinfo}

\alias{genotypesFromPinfo}

\alias{genotypesFromPinfo}

\alias{genotypesFromPinfo}

\alias{genotypesFromPinfo}

