\name{getRefBase}
\alias{getRefBase}
\title{Look up reference base.}
\usage{getRefBase(gs)
getRefBase(gs)
getRefBase(gs)
getRefBase(gs)
getRefBase(gs)
getRefBase(gs)
}
\description{Look up reference base.}
\details{Given a RangedData, look up refrence base at each position.\code{getRefBase}: Given a RangedData, look up refrence base at each position.


\code{getRefBase}: Given a RangedData, look up refrence base at each position.


\code{getRefBase}: Given a RangedData, look up refrence base at each position.


\code{getRefBase}: Given a RangedData, look up refrence base at each position.


\code{getRefBase}: Given a RangedData, look up refrence base at each position.


}
\value{character, nucleotide at each position of rd in reference genome\code{getRefBase}: character, nucleotide at each position of rd in reference genome


\code{getRefBase}: character, nucleotide at each position of rd in reference genome


\code{getRefBase}: character, nucleotide at each position of rd in reference genome


\code{getRefBase}: character, nucleotide at each position of rd in reference genome


\code{getRefBase}: character, nucleotide at each position of rd in reference genome


}
\author{Peter M. Haverty \email{phaverty@gene.com}}
\arguments{\item{gs}{Genoset, RangedData, or RangesList}
}

\alias{getRefBase}

\alias{getRefBase}

\alias{getRefBase}

\alias{getRefBase}

\alias{getRefBase}

