% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNOMAD_functions.R
\name{gnomad_coding_protein_MAF}
\alias{gnomad_coding_protein_MAF}
\title{Title}
\usage{
gnomad_coding_protein_MAF(chr, codingstr, proteinstr, position, gnomad)
}
\arguments{
\item{gnomad}{}
}
\description{
Title
}
