% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNOMAD_functions.R
\name{gnomad_ref_alt_MAF}
\alias{gnomad_ref_alt_MAF}
\title{Extract Minor Allel Frequency from gnomad table}
\usage{
gnomad_ref_alt_MAF(chr, ref, alt, position, gnomad)
}
\arguments{
\item{gnomad}{}
}
\description{
Extract Minor Allel Frequency from gnomad table
}
