\name{headerLines}
\alias{headerLines}
\title{Format list as VCF header lines...}
\usage{headerLines(x, type)
headerLines(x, type)
headerLines(x, type)
headerLines(x, type)
headerLines(x, type)
headerLines(x, type)
}
\description{Format list as VCF header lines}
\details{Format list as VCF header lines\code{headerLines}: Format list as VCF header lines


\code{headerLines}: Format list as VCF header lines


\code{headerLines}: Format list as VCF header lines


\code{headerLines}: Format list as VCF header lines


\code{headerLines}: Format list as VCF header lines


}
\value{list of characters\code{headerLines}: list of characters


\code{headerLines}: list of characters


\code{headerLines}: list of characters


\code{headerLines}: list of characters


\code{headerLines}: list of characters


}
\author{Peter M. Haverty \email{phaverty@gene.com}}
\arguments{\item{x}{list of lists like info.list, format.list, etc. in writeVCFHeader}
\item{type}{character, like INFO or FORMAT or FILTER}
}

\alias{headerLines}

\alias{headerLines}

\alias{headerLines}

\alias{headerLines}

\alias{headerLines}

