% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HORAK_function_call.R
\name{Horak_classification}
\alias{Horak_classification}
\title{Assign Horak Score classifications to values}
\usage{
Horak_classification(HorakScore)
}
\arguments{
\item{HorakScore}{}
}
\description{
Assign Horak Score classifications to values
}
