% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HORAK_function_call.R
\name{Horak_score_cancerHotspot_counts}
\alias{Horak_score_cancerHotspot_counts}
\title{Apply Horak scoring rule to considering cancerHotspot mutation counts}
\usage{
Horak_score_cancerHotspot_counts(cancerhotspotpath)
}
\arguments{
\item{cancerhotspotpath}{}
}
\description{
Apply Horak scoring rule to considering cancerHotspot mutation counts
}
