% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HORAK_function_call.R
\name{Horak_score_function_calls}
\alias{Horak_score_function_calls}
\title{Extract paths for annotation tables and call Horak scoring functions}
\usage{
Horak_score_function_calls(annotation_paths)
}
\arguments{
\item{annotation_paths}{}
}
\description{
Extract paths for annotation tables and call Horak scoring functions
}
