% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HORAK_function_call.R
\name{Horak_score_gnomad}
\alias{Horak_score_gnomad}
\title{Apply Horak scoring rule to gnomad MAF
if MAF > 5\%, -8
if 1\% < MAF < 5\%, -4
absent or < 1\%, 1}
\usage{
Horak_score_gnomad(gnomadpath)
}
\arguments{
\item{gnomadpath}{}
}
\description{
Apply Horak scoring rule to gnomad MAF
if MAF > 5\%, -8
if 1\% < MAF < 5\%, -4
absent or < 1\%, 1
}
