% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HORAK_function_call.R
\name{Horak_score_oncogenpos}
\alias{Horak_score_oncogenpos}
\title{Apply Horak scoring rule to considering cancerHotspot mutation counts}
\usage{
Horak_score_oncogenpos(oncogenPos_path)
}
\arguments{
\item{oncogenPos_path}{}
}
\description{
Apply Horak scoring rule to considering cancerHotspot mutation counts
}
