% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HORAK_function_call.R
\name{HorakScore}
\alias{HorakScore}
\title{Sum up individual module scores to generate Horak Score}
\usage{
HorakScore(horak_scores, horakScore_fp)
}
\arguments{
\item{horak_scores}{}
}
\description{
Sum up individual module scores to generate Horak Score
}
