% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRR_gene_check.R
\name{HRR_check}
\alias{HRR_check}
\title{Check if gene is in HRR gene list}
\usage{
HRR_check(genestr, mp_vars_table)
}
\arguments{
\item{mp_vars_table}{}
}
\description{
Check if gene is in HRR gene list
}
