% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRR_gene_check.R
\name{HRR_check_retrieve_table}
\alias{HRR_check_retrieve_table}
\title{HRR gene check table retrieve}
\usage{
HRR_check_retrieve_table(variant_table, mp_vars_table)
}
\arguments{
\item{mp_vars_table}{}
}
\description{
HRR gene check table retrieve
}
