% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_CNVs.R
\name{hrr_cnv}
\alias{hrr_cnv}
\title{Indicate if gene is present in HRR gene list}
\usage{
hrr_cnv(gene, hrrgenes)
}
\arguments{
\item{hrrgenes}{}
}
\description{
Indicate if gene is present in HRR gene list
}
