% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\name{Insertion}
\alias{Insertion}
\title{Insertion constructor}
\usage{
Insertion(variantID, position, altAllele, ...)
}
\arguments{
\item{variantID}{Variant identifier}

\item{position}{GRanges object (insertion location)}

\item{altAllele}{Inserted sequence}

\item{...}{Additional Variant slots}
}
\value{
Insertion object
}
\description{
Creates an Insertion object
}



\examples{
# Create a GRanges object for a range
pos_ins <- GRanges(seqnames = "chr17",
                   ranges = IRanges(start = 41276045, end = 41276045),
                   strand = "+")
#crate an insertion
ins <- Insertion(
  variantID = "COSM476",
  position = pos_ins,
  altAllele = "AG",
  geneSymbol = "BRCA1",
  geneType = "protein-coding",
  effect = "frameshift_variant",
  clinicalRelevance = "Pathogenic",
  ensemblID = "ENST00000357654",
  hgvsp = "p.S1615fs"
)

ins
}
