\name{loadPinfo}
\alias{loadPinfo}
\title{Read probeset info file...}
\usage{loadPinfo(pinfo.file, mapping="hg19")
loadPinfo(pinfo.file, mapping="hg19")
loadPinfo(pinfo.file, mapping="hg19")
loadPinfo(pinfo.file, mapping="hg19")
loadPinfo(pinfo.file, mapping="hg19")
loadPinfo(pinfo.file, mapping="hg19")
}
\description{Read probeset info file}
\details{Read file from Illumina with probe locs and possible nucleotides\code{loadPinfo}: Read file from Illumina with probe locs and possible nucleotides


\code{loadPinfo}: Read file from Illumina with probe locs and possible nucleotides


\code{loadPinfo}: Read file from Illumina with probe locs and possible nucleotides


\code{loadPinfo}: Read file from Illumina with probe locs and possible nucleotides


\code{loadPinfo}: Read file from Illumina with probe locs and possible nucleotides


}
\value{data.frame of pinfo.file contents with a little tidying up.\code{loadPinfo}: data.frame of pinfo.file contents with a little tidying up.


\code{loadPinfo}: data.frame of pinfo.file contents with a little tidying up.


\code{loadPinfo}: data.frame of pinfo.file contents with a little tidying up.


\code{loadPinfo}: data.frame of pinfo.file contents with a little tidying up.


\code{loadPinfo}: data.frame of pinfo.file contents with a little tidying up.


}
\author{Peter M. Haverty \email{phaverty@gene.com}}
\arguments{\item{pinfo.file}{character, name of file from Illumina with probe location and alleles}
\item{mapping}{character, genome version (e.g. hg19)}
}

\alias{loadPinfo}

\alias{loadPinfo}

\alias{loadPinfo}

\alias{loadPinfo}

\alias{loadPinfo}

