\name{makeAltColumn}
\alias{makeAltColumn}
\title{Make ALT column...}
\usage{makeAltColumn(allele.lookup)
makeAltColumn(allele.lookup)
makeAltColumn(allele.lookup)
makeAltColumn(allele.lookup)
makeAltColumn(allele.lookup)
makeAltColumn(allele.lookup)
}
\description{Make ALT column}
\details{Take lookup of allele options from codeAlleleOptions() and output ALT column
string vector.  Basically take out ref option and use "." for only ref option.
This contains the sloweset loop ever. Someone please think of a way to do this
without mclapply.\code{makeAltColumn}: Take lookup of allele options from codeAlleleOptions() and output ALT column
string vector.  Basically take out ref option and use "." for only ref option.
This contains the sloweset loop ever. Someone please think of a way to do this
without mclapply.


\code{makeAltColumn}: Take lookup of allele options from codeAlleleOptions() and output ALT column
string vector.  Basically take out ref option and use "." for only ref option.
This contains the sloweset loop ever. Someone please think of a way to do this
without mclapply.


\code{makeAltColumn}: Take lookup of allele options from codeAlleleOptions() and output ALT column
string vector.  Basically take out ref option and use "." for only ref option.
This contains the sloweset loop ever. Someone please think of a way to do this
without mclapply.


\code{makeAltColumn}: Take lookup of allele options from codeAlleleOptions() and output ALT column
string vector.  Basically take out ref option and use "." for only ref option.
This contains the sloweset loop ever. Someone please think of a way to do this
without mclapply.


\code{makeAltColumn}: Take lookup of allele options from codeAlleleOptions() and output ALT column
string vector.  Basically take out ref option and use "." for only ref option.
This contains the sloweset loop ever. Someone please think of a way to do this
without mclapply.


}
\value{character, comma-separated list of alelle options, "." if alt is only ref\code{makeAltColumn}: character, comma-separated list of alelle options, "." if alt is only ref


\code{makeAltColumn}: character, comma-separated list of alelle options, "." if alt is only ref


\code{makeAltColumn}: character, comma-separated list of alelle options, "." if alt is only ref


\code{makeAltColumn}: character, comma-separated list of alelle options, "." if alt is only ref


\code{makeAltColumn}: character, comma-separated list of alelle options, "." if alt is only ref


}
\author{Peter M. Haverty email{phaverty@gene.com}}
\arguments{\item{allele.lookup}{list, like list( c("A","C"), c("G"), c("T","A","G"))}
}

\alias{makeAltColumn}

\alias{makeAltColumn}

\alias{makeAltColumn}

\alias{makeAltColumn}

\alias{makeAltColumn}

