% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_variant_table.R
\name{MP_check_retrieve_table}
\alias{MP_check_retrieve_table}
\title{Apply MP_variant_check at table level}
\usage{
MP_check_retrieve_table(variant_table, mp_vars_table)
}
\arguments{
\item{mp_vars_table}{}
}
\description{
Apply MP_variant_check at table level
}
