% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_variant_table.R
\name{MP_variant_check}
\alias{MP_variant_check}
\title{Check MP_variant_table for variant interpretations}
\usage{
MP_variant_check(rowid, genestr, codingstr, protein, mp_vars_table)
}
\arguments{
\item{mp_vars_table}{}
}
\description{
Check MP_variant_table for variant interpretations
}
