% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_counter.R
\name{mutation_amino_acid}
\alias{mutation_amino_acid}
\title{Extract mutation amino acid from 'appropriate' variant entries}
\usage{
mutation_amino_acid(aa_change)
}
\arguments{
\item{aa_change}{}
}
\description{
Extract mutation amino acid from 'appropriate' variant entries
}
