% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncogen_pos_function_call.R
\name{oncogen_pos_retrieve_interpretation}
\alias{oncogen_pos_retrieve_interpretation}
\title{Retrieve if variant has been identified as oncogenic position}
\usage{
oncogen_pos_retrieve_interpretation(
  gene,
  coding_change,
  protein_change,
  oncogen_pos
)
}
\arguments{
\item{oncogen_pos}{}
}
\description{
Retrieve if variant has been identified as oncogenic position
}
