% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncogen_pos_function_call.R
\name{oncogenic_Position}
\alias{oncogenic_Position}
\title{Retrieve if variant has been identified as oncogenic position}
\usage{
oncogenic_Position(geneid, ChangePosition, oncogen_vars)
}
\arguments{
\item{oncogen_pos}{}
}
\description{
Retrieve if variant has been identified as oncogenic position
}
