% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\docType{class}
\name{ONV-class}
\alias{ONV-class}
\title{Other Nucleotide Variant (ONV) class}
\description{
Class for complex nucleotide changes
}
\section{Slots}{

\describe{
\item{\code{variantType}}{Type of variant}
}}

