% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_vep_api.R
\name{post_Hs_region}
\alias{post_Hs_region}
\title{elementary vep/homo_sapiens/region call to ensembl VEP REST API}
\usage{
post_Hs_region(chr, pos, id, ref, alt)
}
\arguments{
\item{chr}{character(1) ensembl chromosome identifier (e.g., "7")}

\item{pos}{numeric(1) 1-based chromosome position}

\item{id}{character(1) arbitrary identifier}

\item{ref}{character(1) reference allele}

\item{alt}{character(1) alternative allele}
}
\value{
Instance of 'response' defined in httr package.
}
\description{
elementary vep/homo_sapiens/region call to ensembl VEP REST API
}
\note{
This function prepares a POST to rest.ensembl.org/vep/homo_sapiens/region endpoint.
}
\examples{
chk = post_Hs_region("7", 155800001, "chk", "A", "T")
chk
res = jsonlite::fromJSON(jsonlite::toJSON(httr::content(chk)))
dim(chk)
}
