% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_counter.R
\name{ref_amino_acid}
\alias{ref_amino_acid}
\title{Extract reference amino acid from 'appropriate' variant entries}
\usage{
ref_amino_acid(aa_change)
}
\arguments{
\item{aa_change}{}
}
\description{
Extract reference amino acid from 'appropriate' variant entries
}
