% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\docType{class}
\name{SNV-class}
\alias{SNV-class}
\title{Single Nucleotide Variant (SNV) class}
\description{
Class for single nucleotide substitutions (e.g., A>T)
}


\examples{
# Create a dummy Variant object (assuming class definition exists)
pos <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, width=1), strand = "+")
snv <- SNV(
  variantID = "rs1",
  position = pos,
  refAllele = "A",
  altAllele = "T",
  geneSymbol = "GENE1",
  geneType = "protein-coding",
  effect = "missense",
  clinicalRelevance = "Pathogenic",
  ensemblID = "ENST0000001",
  hgvsp = "p.Val1Gly"
)

snv
}

