% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\name{SNV}
\alias{SNV}
\title{SNV constructor}
\usage{
SNV(variantID, position, refAllele, altAllele, ...)
}
\arguments{
\item{variantID}{Variant identifier}

\item{position}{GRanges object with position}

\item{refAllele}{Reference allele (single nucleotide)}

\item{altAllele}{Alternate allele (single nucleotide)}

\item{...}{Additional Variant slots}
}
\value{
SNV object
}
\description{
Creates an SNV (single nucleotide variant) object
}

\examples{
# Create a dummy Variant object (assuming class definition exists)
pos <- GenomicRanges::GRanges("chr1", IRanges::IRanges(100, width=1), strand = "+")
snv <- SNV(
  variantID = "rs1",
  position = pos,
  refAllele = "A",
  altAllele = "T",
  geneSymbol = "GENE1",
  geneType = "protein-coding",
  effect = "missense",
  clinicalRelevance = "Pathogenic",
  ensemblID = "ENST0000001",
  hgvsp = "p.Val1Gly"
)

snv
}

