% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLINVAR_functions.R
\name{somatic_filtering}
\alias{somatic_filtering}
\title{Filter for 'somatic' classification  in case of multiple ClinVar entries}
\usage{
somatic_filtering(clinvarhits)
}
\arguments{
\item{clinvarhits}{}
}
\description{
Filter for 'somatic' classification  in case of multiple ClinVar entries
}
