% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncogen_pos_function_call.R
\name{table_retrieve_oncogen_positions}
\alias{table_retrieve_oncogen_positions}
\title{Apply oncogenic position check to whole table.}
\usage{
table_retrieve_oncogen_positions(filepath, oncogen_pos)
}
\arguments{
\item{oncogen_pos}{}
}
\description{
Apply oncogenic position check to whole table.
}
