% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oncogen_pos_function_call.R
\name{table_retrieve_oncogenic_information}
\alias{table_retrieve_oncogenic_information}
\title{Apply oncogenic position check to whole table.}
\usage{
table_retrieve_oncogenic_information(filepath, oncogen_vars)
}
\arguments{
\item{oncogen_pos}{}
}
\description{
Apply oncogenic position check to whole table.
}
