% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_output.R
\name{tech_class}
\alias{tech_class}
\title{\enumerate{
\item Sheet -- Technical classification of variants, e.g. if artefacts or not
}}
\usage{
tech_class(snvdf)

tech_class(snvdf)
}
\arguments{
\item{snvdf}{}
}
\description{
\enumerate{
\item Sheet -- Technical classification of variants, e.g. if artefacts or not
}

\enumerate{
\item Sheet -- Interpretation tools
}
}
