% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTSGpos.R
\name{TSG_check_function_call}
\alias{TSG_check_function_call}
\title{tsgParseTable Checks SNV contains 'fs' or 'Ter'}
\usage{
TSG_check_function_call(snvtable, TSG_list)
}
\arguments{
\item{snvtable}{}
}
\value{
snvtable with tsgInfo column added
}
\description{
tsgParseTable Checks SNV contains 'fs' or 'Ter'
}
