% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_CNVs.R
\name{tsg_cnv_effect}
\alias{tsg_cnv_effect}
\title{Determine if CNV loss is happening in TSG}
\usage{
tsg_cnv_effect(gene, cnv_status, tsg_list)
}
\arguments{
\item{tsg_list}{}
}
\description{
Determine if CNV loss is happening in TSG
}
