% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_variants.R
\docType{class}
\name{Variant-class}
\alias{Variant-class}
\title{Virtual general class for all variant types}
\description{
This class serves as the parent for all specific variant types.
}
\section{Slots}{

\describe{
\item{\code{variantID}}{Unique identifier for the variant (e.g., "rs123")}

\item{\code{position}}{GRanges object with genomic position}

\item{\code{refAllele}}{Reference allele sequence}

\item{\code{altAllele}}{Alternate allele sequence}

\item{\code{geneSymbol}}{Gene symbol affected (e.g., "BRCA1")}

\item{\code{geneType}}{Type of gene ("protein-coding", "lncRNA", etc.)}

\item{\code{effect}}{Variant effect ("missense", "nonsense", etc.)}

\item{\code{clinicalRelevance}}{Clinical significance from ClinVar/COSMIC}

\item{\code{ensemblID}}{Ensembl transcript ID}

\item{\code{hgvsp}}{HGVSp protein notation (e.g., "p.Arg123Leu")}
}}

