% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_vep_api.R
\name{variant_body}
\alias{variant_body}
\title{helper function to construct inputs for VEP REST API}
\usage{
variant_body(chr, pos, id, ref, alt)
}
\arguments{
\item{chr}{character(1)}

\item{pos}{numeric(1)}

\item{id}{character(1)}

\item{ref}{character(1)}

\item{alt}{character(1)}
}
\description{
helper function to construct inputs for VEP REST API
}
\note{
Produces a string used as an example in VEP API documentation.
}
