\name{VAUtil-class}
\docType{class}
\alias{.SRUtil-class}


\title{".SRUtil" and potentially other related classes}
\description{
  These classes provide important utility functions in the
  \pkg{VariantAnnotation} package, but may occasionally be seen by the user and
  are documented here for that reason.
}
\section{Objects from the Class}{

  Utility classes include:

  \itemize{
    \item \code{.SRUtil-class} a virtual base class from which all
    utility classes are derived.
  }

}

\author{Valerie Obenchain}

\examples{
getClass(".SRUtil", where=getNamespace("VariantAnnotation"))
}

\keyword{classes}
