% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationModules.R
\name{write_Annotation_Modules}
\alias{write_Annotation_Modules}
\title{Calls Annotation Module functions for cancerHotspots, COSMIC, GNOMAD, ClinVar}
\usage{
write_Annotation_Modules(snvt, annotation_fp)
}
\arguments{
\item{annotation_fp}{}
}
\description{
Calls Annotation Module functions for cancerHotspots, COSMIC, GNOMAD, ClinVar
}
