% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerHotspots_counter.R
\name{write_out_cancerHotspot_info}
\alias{write_out_cancerHotspot_info}
\title{Title}
\usage{
write_out_cancerHotspot_info(CH_snv, CH_info_file)
}
\arguments{
\item{CH_info_file}{}
}
\description{
Title
}
