% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSMIC_functions.R
\name{write_out_cosmic_info}
\alias{write_out_cosmic_info}
\title{Write out cancer Hotspot info table}
\usage{
write_out_cosmic_info(COSMIC_snv, COSMIC_info_file)
}
\arguments{
\item{CH_info_file}{}
}
\description{
Write out cancer Hotspot info table
}
