\name{writeVCFHeader}
\alias{writeVCFHeader}
\title{Write header part of VCF...}
\usage{writeVCFHeader(filename, info.list, filter.list, format.list, sample.names)
writeVCFHeader(filename, info.list, filter.list, format.list, sample.names)
writeVCFHeader(filename, info.list, filter.list, format.list, sample.names)
writeVCFHeader(filename, info.list, filter.list, format.list, sample.names)
writeVCFHeader(filename, info.list, filter.list, format.list, sample.names)
writeVCFHeader(filename, info.list, filter.list, format.list, sample.names)
}
\description{Write header part of VCF}
\details{Write header part of VCF\code{writeVCFHeader}: Write header part of VCF


\code{writeVCFHeader}: Write header part of VCF


\code{writeVCFHeader}: Write header part of VCF


\code{writeVCFHeader}: Write header part of VCF


\code{writeVCFHeader}: Write header part of VCF


}
\value{nothing\code{writeVCFHeader}: nothing


\code{writeVCFHeader}: nothing


\code{writeVCFHeader}: nothing


\code{writeVCFHeader}: nothing


\code{writeVCFHeader}: nothing


}
\author{Peter M. Haverty \email{phaverty@gene.com}}
\arguments{\item{filename}{character, name of file to create and write to}
\item{info.list}{list of lists. Sub-lists are name/value pairs for INFO rows.}
\item{filter.list}{list of lists. Sub-lists are name/value pairs for FILTER rows.}
\item{format.list}{list of lists. Sub-lists are name/value pairs for FORMAT rows.}
\item{sample.names}{character, names of sample to go in last header row}
}

\alias{writeVCFHeader}

\alias{writeVCFHeader}

\alias{writeVCFHeader}

\alias{writeVCFHeader}

\alias{writeVCFHeader}

