\name{variants}
\alias{variants}
\docType{data}

\title{
  Human genetic variant test dataset
}

\description{
  This dataset contains fabricated Human genetic variants. 
}

\usage{data(variants)}

\format{
  The dataset is an object of class \code{GRanges}. Information for each 
  variant includes chromosome, range, strand, reference and observed allele, 
  and a comment.
}

\examples{
data(variants)
variants
}
\keyword{datasets}
