% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_output.R
\name{variants}
\alias{variants}
\title{\enumerate{
\item Sheet -- Overview of detected variantas
}}
\usage{
variants(snvdf)
}
\arguments{
\item{snvdf}{}
}
\description{
\enumerate{
\item Sheet -- Overview of detected variantas
}
}
